+++ dart/lib/src/hub_adapter.dart
@@ -8,6 +8,7 @@ import 'profiling.dart';
 import 'protocol.dart';
 import 'protocol/sentry_feedback.dart';
 import 'scope.dart';
+import 'sdk_lifecycle_hooks.dart';
 import 'sentry.dart';
 import 'sentry_client.dart';
 import 'sentry_options.dart';
@@ -201,8 +202,18 @@ class HubAdapter implements Hub {
   FutureOr<void> captureLog(SentryLog log) => Sentry.currentHub.captureLog(log);
 
   @override
-  void registerCallback<T extends SdkLifecycleEvent>(
+  Map<Type, List<Function>> get lifecycleCallbacks =>
+      Sentry.currentHub.lifecycleCallbacks;
+
+  @override
+  void registerSdkLifecycleCallback<T extends SdkLifecycleEvent>(
       SdkLifecycleCallback<T> callback) {
-    Sentry.currentHub.registerCallback(callback);
+    Sentry.currentHub.registerSdkLifecycleCallback(callback);
+  }
+
+  @override
+  void removeSdkLifecycleCallback<T extends SdkLifecycleEvent>(
+      SdkLifecycleCallback<T> callback) {
+    Sentry.currentHub.removeSdkLifecycleCallback(callback);
   }
 }